feat: homepage completed
