+++ src/components/homepage/EventsSection.tsx
@@ -17,7 +17,6 @@ const EventsSection: Component = () => {
 
   return (
     <section className="relative overflow-hidden bg-slate-900 py-20">
-      {/* Background Elements */}
       <div className="absolute inset-0">
         <div className="absolute left-10 top-20 size-32 rounded-full bg-green-500/5 blur-3xl" />
         <div className="absolute bottom-20 right-10 size-40 rounded-full bg-emerald-500/5 blur-3xl" />
@@ -25,7 +24,6 @@ const EventsSection: Component = () => {
 
       <div className="container relative z-10 mx-auto px-6">
         <div className="mx-auto max-w-6xl">
-          {/* Header */}
           <ScrollReveal>
             <div className="mb-12 text-center">
               <h2 className="mb-6 text-4xl font-bold lg:text-5xl">
@@ -44,7 +42,6 @@ const EventsSection: Component = () => {
             </div>
           </ScrollReveal>
 
-          {/* Tabs */}
           <ScrollReveal delay={200}>
             <div className="mb-12 flex justify-center">
               <div className="rounded-lg border border-gray-700 bg-slate-800 p-1">
@@ -74,7 +71,6 @@ const EventsSection: Component = () => {
             </div>
           </ScrollReveal>
 
-          {/* Events Grid */}
           <div className="mb-12 grid gap-8 md:grid-cols-2 lg:grid-cols-3">
             {activeTab === 'upcoming' &&
               upcomingEvents.map((event, index) => (
@@ -91,13 +87,12 @@ const EventsSection: Component = () => {
               ))}
           </div>
 
-          {/* CTA Section */}
           <ScrollReveal>
             <div className="rounded-2xl border border-gray-700 bg-gradient-to-r from-slate-800/50 to-slate-700/50 p-8 text-center backdrop-blur-sm">
               <h3 className="mb-4 text-3xl font-bold text-white">Want to Speak at Our Events?</h3>
               <p className="mx-auto mb-6 max-w-2xl text-lg text-gray-300">
-                Share your knowledge and experience with our community. We're always looking for
-                passionate speakers.
+                Share your knowledge and experience with our community. We&apos;re always looking
+                for passionate speakers.
               </p>
               <div className="flex flex-col justify-center gap-4 sm:flex-row">
                 <Link
